+++ src/extension.ts
@@ -2,9 +2,13 @@
 
 import * as vscode from 'vscode';
 import * as request from 'request';
+import * as url from 'url';
+import opn = require('opn');
+import { LDStreamProcessor, LDFlagValue, LDOptions } from 'ldclient-node';
 import InMemoryFeatureStore = require('ldclient-node/feature_store');
 import StreamProcessor = require('ldclient-node/streaming');
 import Requestor = require('ldclient-node/requestor');
+import { kebabCase } from 'lodash';
 
 const LD_MODE: vscode.DocumentFilter = {
 	scheme: 'file',
@@ -12,8 +16,9 @@ const LD_MODE: vscode.DocumentFilter = {
 const DATA_KIND = { namespace: 'features' };
 const FLAG_KEY_REGEX = /[A-Za-z0-9][\.A-Za-z_\-0-9]*/;
 
-var store;
-var updateProcessor;
+//TODO: update LDFeatureStore type to support versioned data kind
+let store: any;
+let updateProcessor: LDStreamProcessor;
 
 class LaunchDarklyCompletionItemProvider implements vscode.CompletionItemProvider {
 	public provideCompletionItems(
@@ -40,8 +45,7 @@ class LaunchDarklyHoverProvider implements vscode.HoverProvider {
 		token: vscode.CancellationToken,
 	): Thenable<vscode.Hover> {
 		return new Promise(resolve => {
-			store.all(DATA_KIND, flags => {
-				var flag = flags[document.getText(document.getWordRangeAtPosition(position, FLAG_KEY_REGEX))];
+			getFlagKeyAtCurrentPosition(document, position, flag => {
 				if (flag) {
 					let hoverString = `**LaunchDarkly Feature Flag**\n
 Key: ${flag.key}\n
@@ -53,24 +57,25 @@ Prerequisite count: ${flag.prerequisites.length}\n
 Target count: ${flag.targets.length}\n
 Rule count: ${flag.rules.length}`;
 					resolve(new vscode.Hover(hoverString));
-				} else {
-					resolve();
+					return;
 				}
+				resolve();
 			});
 		});
 	}
 }
 
 export function activate(ctx: vscode.ExtensionContext) {
-	const channel = vscode.window.createOutputChannel('LaunchDarkly');
-	var settings = vscode.workspace.getConfiguration('launchdarkly');
-	if (settings.get('sdkKey')) {
-		var sdkKey = settings.get('sdkKey');
+	let settings = vscode.workspace.getConfiguration('launchdarkly');
+	let sdkKey = settings.get<string>('sdkKey');
+	if (sdkKey) {
+		let baseUri = settings.get<string>('baseUri');
+		let streamUri = settings.get<string>('streamUri');
 		store = InMemoryFeatureStore();
-		var config = {
+		let config: LDOptions = {
 			timeout: 5,
-			baseUri: 'https://app.launchdarkly.com',
-			streamUri: 'https://stream.launchdarkly.com',
+			baseUri: baseUri,
+			streamUri: streamUri,
 			featureStore: store,
 			// noop logger for debug calls
 			logger: { debug: () => {} },
@@ -78,6 +83,10 @@ export function activate(ctx: vscode.ExtensionContext) {
 		updateProcessor = StreamProcessor(sdkKey, config, Requestor(sdkKey, config));
 		updateProcessor.start(function(err) {
 			if (err) {
+				let errMsg = `[LaunchDarkly] Error retrieving flags.${baseUri != 'https://app.launchdarkly.com' ||
+				streamUri != 'https://stream.launchdarkly.com'
+					? ' Please make sure your configured base and stream URIs are correct'
+					: ''}`;
 				vscode.window.showErrorMessage('[LaunchDarkly] Error retrieving flags.');
 			} else {
 				process.nextTick(function() {
@@ -98,17 +107,15 @@ export function activate(ctx: vscode.ExtensionContext) {
 	}
 
 	ctx.subscriptions.push(
-		vscode.commands.registerTextEditorCommand('extension.getFlag', () => {
-			let flag = getFlagFromEditorSelection();
-			if (flag === '') {
-				vscode.window.showErrorMessage('[LaunchDarkly] Error retrieving flag (no selection made).');
+		vscode.commands.registerTextEditorCommand('extension.openInLaunchDarkly', editor => {
+			let flagKey = editor.document.getText(
+				editor.document.getWordRangeAtPosition(editor.selection.anchor, FLAG_KEY_REGEX),
+			);
+			if (flagKey === '') {
+				vscode.window.showErrorMessage(
+					'[LaunchDarkly] Error retrieving flag (current cursor position is not a feature flag).',
+				);
 				return;
-			}
-
-			if (settings.get('clearOutputBeforeEveryCommand')) {
-				channel.clear();
-			} else {
-				channel.appendLine('Getting flag: ' + flag);
 			}
 
 			if (!settings.get('accessToken')) {
@@ -117,27 +124,19 @@ export function activate(ctx: vscode.ExtensionContext) {
 			}
 
 			let project = getProject(settings);
-			let envParam = '?env=' + getEnvironment(settings);
+			if (!project) {
+				vscode.window.showErrorMessage('[LaunchDarkly] project is not set.');
+				return;
+			}
 
-			var options = {
-				url: 'https://app.launchdarkly.com/api/v2/flags/' + project + '/' + flag + envParam,
-				headers: {
-					Authorization: settings.get('accessToken'),
-				},
-			};
-
-			request(options, (error, response, body) => {
-				if (!error) {
-					if (response.statusCode == 200) {
-						channel.appendLine(JSON.stringify(JSON.parse(body), null, 2));
-						channel.show();
-					} else if (response.statusCode == 404) {
-						vscode.window.showErrorMessage('[LaunchDarkly] 404 - Flag not found.');
-					} else {
-						vscode.window.showErrorMessage(response.statusCode);
-					}
+			getFeatureFlag(settings, flagKey, (flag: LDFlagValue) => {
+				console.log(flag)
+				let baseUri = settings.get<string>('baseUri');
+				let env = getEnvironment(settings);
+				if (env === '') {
+					opn(url.resolve(baseUri, flag.environments[Object.keys(flag.environments)[0]]._site.href));
 				} else {
-					vscode.window.showErrorMessage('[LaunchDarkly] Encountered an error retrieving flag.');
+					opn(url.resolve(baseUri, flag.environments[env]._site.href));
 				}
 			});
 		}),
@@ -148,28 +147,65 @@ export function deactivate() {
 	updateProcessor.stop();
 }
 
-function getFlagFromEditorSelection() {
-	let editor = vscode.window.activeTextEditor;
-	let selection = editor.selection;
-	return editor.document.getText(
-		new vscode.Range(selection.start.line, selection.start.character, selection.end.line, selection.end.character),
-	);
+function getProject(settings: vscode.WorkspaceConfiguration) {
+	return settings.get<string>('project');
 }
 
-function getProject(settings) {
-	if (settings.get('project')) {
-		return settings.get('project');
-	} else {
-		vscode.window.showWarningMessage("[LaunchDarkly] project is not set, using 'default' project instead.");
-		return 'default';
+function getEnvironment(settings: vscode.WorkspaceConfiguration) {
+	if (settings.get<string>('env')) {
+		return settings.get<string>('env');
 	}
+	vscode.window.showWarningMessage('[LaunchDarkly] env is not set. Falling back to first environment.');
+	return '';
 }
 
-function getEnvironment(settings) {
-	if (settings.get('env')) {
-		return settings.get('env');
-	} else {
-		vscode.window.showInformationMessage('[LaunchDarkly] env is not set. Showing all environments.');
-		return '';
-	}
+function getFlagKeyAtCurrentPosition(document: vscode.TextDocument, position: vscode.Position, cb: Function) {
+	store.all(DATA_KIND, flags => {
+		let candidate = document.getText(document.getWordRangeAtPosition(position, FLAG_KEY_REGEX));
+		cb(flags[candidate] || flags[kebabCase(candidate)]);
+	});
+}
+
+function getFeatureFlag(settings: vscode.WorkspaceConfiguration, flagKey: string, cb: Function) {
+	let baseUri = settings.get<string>('baseUri');
+	let project = getProject(settings);
+	let env = getEnvironment(settings);
+	let envParam = env ? '?env=' + env : '';
+	let options = {
+		url: url.resolve(baseUri, `api/v2/flags/${project}/${flagKey + envParam}`),
+		headers: {
+			Authorization: settings.get('accessToken'),
+		},
+	};
+	request(options, (error, response, body) => {
+		if (!error) {
+			if (response.statusCode == 200) {
+				cb(JSON.parse(body).environments);
+			} else if (response.statusCode == 404) {
+				// Try resolving the flag key to kebab case
+				options.url = url.resolve(baseUri, `api/v2/flags/${project}/${kebabCase(flagKey) + envParam}`);
+				console.log(options)
+				request(options, (error, response, body) => {
+					if (!error) {
+						if (response.statusCode == 200) {
+							console.log(JSON.parse(body).environment)
+							cb(JSON.parse(body).environments);
+						} else if (response.statusCode == 404) {
+							vscode.window.showErrorMessage(`[LaunchDarkly] Flag ${flagKey} not found.`);
+							return;
+						} else {
+							vscode.window.showErrorMessage(response.statusCode);
+						}
+					} else {
+						vscode.window.showErrorMessage(`[LaunchDarkly] Encountered an error retrieving the flag ${flagKey}`);
+					}
+				});
+			} else {
+				vscode.window.showErrorMessage(response.statusCode);
+			}
+		} else {
+			vscode.window.showErrorMessage(`[LaunchDarkly] Encountered an error retrieving the flag ${flagKey}`);
+		}
+		cb(error, response, body);
+	});
 }