+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
+
 All notable changes to the "launchdarkly" extension will be documented in this file.
 
+## [2.0.0] - 2018-08-16
+
+### Added
+
+- The LaunchDarkly base and stream uris are now configurable.
+
+### Changed
+
+- Replace the "Get feature flag" command with an "Open in LaunchDarkly" command to view a feature flag in the LaunchDarkly dashboard.
+
+### Fixed
+
+- Flag keys with non-letter characters are now correctly discovered by the extension
+
 ## [1.0.0] - 2017-08-10
+
 - Initial release